+++ inject/src/main/kotlin/fr/xgouchet/elmyr/inject/DefaultForgeryInjector.kt
@@ -8,6 +8,7 @@ import fr.xgouchet.elmyr.annotation.FloatForgery
 import fr.xgouchet.elmyr.annotation.Forgery
 import fr.xgouchet.elmyr.annotation.IntForgery
 import fr.xgouchet.elmyr.annotation.LongForgery
+import fr.xgouchet.elmyr.annotation.MapForgery
 import fr.xgouchet.elmyr.annotation.RegexForgery
 import fr.xgouchet.elmyr.annotation.StringForgery
 import fr.xgouchet.elmyr.annotation.StringForgeryType
@@ -15,9 +16,12 @@ import fr.xgouchet.elmyr.inject.reflect.invokePrivate
 import java.lang.IllegalStateException
 import kotlin.reflect.KClass
 import kotlin.reflect.KMutableProperty
+import kotlin.reflect.KParameter
 import kotlin.reflect.KProperty
 import kotlin.reflect.KType
+import kotlin.reflect.KTypeParameter
 import kotlin.reflect.KTypeProjection
+import kotlin.reflect.KVisibility
 import kotlin.reflect.full.declaredMembers
 import kotlin.reflect.full.superclasses
 
@@ -70,7 +74,8 @@ class DefaultForgeryInjector : ForgeryInjector {
                         annotation is DoubleForgery ||
                         annotation is StringForgery ||
                         annotation is RegexForgery ||
-                        annotation is AdvancedForgery
+                        annotation is AdvancedForgery ||
+                        annotation is MapForgery
                 }
             }
         when (invalidProperties.size) {
@@ -104,38 +109,42 @@ class DefaultForgeryInjector : ForgeryInjector {
         property: KMutableProperty<*>,
         target: Any
     ) {
+        var value: Any? = null
         for (annotation in property.annotations) {
-            when (annotation) {
-                is Forgery -> processPropertyWithForgery(forge, property, target)
-                is BoolForgery -> processPropertyWithBooleanForgery(forge, annotation, property, target)
-                is IntForgery -> processPropertyWithIntForgery(forge, annotation, property, target)
-                is LongForgery -> processPropertyWithLongForgery(forge, annotation, property, target)
-                is FloatForgery -> processPropertyWithFloatForgery(forge, annotation, property, target)
-                is DoubleForgery -> processPropertyWithDoubleForgery(forge, annotation, property, target)
-                is StringForgery -> processPropertyWithStringForgery(forge, annotation, property, target)
-                is RegexForgery -> processPropertyWithRegexForgery(forge, annotation, property, target)
-                is AdvancedForgery -> processPropertyWithAdvancedForgery(forge, annotation, property, target)
+            value = when (annotation) {
+                is Forgery -> getForgery(forge, property)
+                is BoolForgery -> getBooleanForgery(forge, annotation, property)
+                is IntForgery -> getIntForgery(forge, annotation, property)
+                is LongForgery -> getLongForgery(forge, annotation, property)
+                is FloatForgery -> getFloatForgery(forge, annotation, property)
+                is DoubleForgery -> getDoubleForgery(forge, annotation, property)
+                is StringForgery -> getStringForgery(forge, annotation, property)
+                is RegexForgery -> getRegexForgery(forge, annotation, property)
+                is AdvancedForgery -> getAdvancedForgery(forge, annotation, property)
+                is MapForgery -> getMapForgery(forge, annotation, property)
+                else -> null
             }
+            if (value != null) break
+        }
+
+        if (value != null) {
+            property.setter.invokePrivate(target, value)
         }
     }
 
-    private fun processPropertyWithForgery(
+    private fun getForgery(
         forge: Forge,
-        property: KMutableProperty<*>,
-        target: Any
-    ) {
-        val forgery = forgeProperty(property.returnType, forge)
-            ?: throw ForgeryInjectorException("Unable to forge $property.")
-
-        property.setter.invokePrivate(target, forgery)
+        property: KMutableProperty<*>
+    ): Any? {
+        return getGenericForgery(property.returnType, forge)
+            ?: uninjectable(property)
     }
 
-    private fun processPropertyWithBooleanForgery(
+    private fun getBooleanForgery(
         forge: Forge,
         annotation: BoolForgery,
-        property: KMutableProperty<*>,
-        target: Any
-    ) {
+        property: KMutableProperty<*>
+    ): Any {
         check(annotation.probability >= 0f) {
             "You can only use an BoolForgery with a probability between 0f and 1f"
         }
@@ -143,18 +152,15 @@ class DefaultForgeryInjector : ForgeryInjector {
             "You can only use an BoolForgery with a probability between 0f and 1f"
         }
         val forgery: Forge.() -> Boolean = { aBool(annotation.probability) }
-        property.setter.invokePrivate(
-            target,
-            forgePrimitiveProperty(property.returnType, Boolean::class, forge, forgery)
-        )
+        val result = getPrimitiveForgery(property.returnType, Boolean::class, forge, forgery)
+        return result ?: uninjectable(property)
     }
 
-    private fun processPropertyWithIntForgery(
+    private fun getIntForgery(
         forge: Forge,
         annotation: IntForgery,
-        property: KMutableProperty<*>,
-        target: Any
-    ) {
+        property: KMutableProperty<*>
+    ): Any {
         val forgery: Forge.() -> Int
         if (annotation.standardDeviation >= 0) {
             check(annotation.min == Int.MIN_VALUE) {
@@ -170,18 +176,15 @@ class DefaultForgeryInjector : ForgeryInjector {
             }
             forgery = { anInt(annotation.min, annotation.max) }
         }
-        property.setter.invokePrivate(
-            target,
-            forgePrimitiveProperty(property.returnType, Int::class, forge, forgery)
-        )
+        val result = getPrimitiveForgery(property.returnType, Int::class, forge, forgery)
+        return result ?: uninjectable(property)
     }
 
-    private fun processPropertyWithLongForgery(
+    private fun getLongForgery(
         forge: Forge,
         annotation: LongForgery,
-        property: KMutableProperty<*>,
-        target: Any
-    ) {
+        property: KMutableProperty<*>
+    ): Any {
         val forgery: Forge.() -> Long
         if (annotation.standardDeviation >= 0) {
             check(annotation.min == Long.MIN_VALUE) {
@@ -197,18 +200,15 @@ class DefaultForgeryInjector : ForgeryInjector {
             }
             forgery = { aLong(annotation.min, annotation.max) }
         }
-        property.setter.invokePrivate(
-            target,
-            forgePrimitiveProperty(property.returnType, Long::class, forge, forgery)
-        )
+        val result = getPrimitiveForgery(property.returnType, Long::class, forge, forgery)
+        return result ?: uninjectable(property)
     }
 
-    private fun processPropertyWithFloatForgery(
+    private fun getFloatForgery(
         forge: Forge,
         annotation: FloatForgery,
-        property: KMutableProperty<*>,
-        target: Any
-    ) {
+        property: KMutableProperty<*>
+    ): Any {
         val forgery: Forge.() -> Float
         if (!annotation.standardDeviation.isNaN()) {
             check(annotation.min == -Float.MAX_VALUE) {
@@ -225,18 +225,15 @@ class DefaultForgeryInjector : ForgeryInjector {
             forgery = { aFloat(annotation.min, annotation.max) }
         }
 
-        property.setter.invokePrivate(
-            target,
-            forgePrimitiveProperty(property.returnType, Float::class, forge, forgery)
-        )
+        val result = getPrimitiveForgery(property.returnType, Float::class, forge, forgery)
+        return result ?: uninjectable(property)
     }
 
-    private fun processPropertyWithDoubleForgery(
+    private fun getDoubleForgery(
         forge: Forge,
         annotation: DoubleForgery,
-        property: KMutableProperty<*>,
-        target: Any
-    ) {
+        property: KMutableProperty<*>
+    ): Any {
         val forgery: Forge.() -> Double
         if (!annotation.standardDeviation.isNaN()) {
             check(annotation.min == -Double.MAX_VALUE) {
@@ -253,18 +250,15 @@ class DefaultForgeryInjector : ForgeryInjector {
             forgery = { aDouble(annotation.min, annotation.max) }
         }
 
-        property.setter.invokePrivate(
-            target,
-            forgePrimitiveProperty(property.returnType, Double::class, forge, forgery)
-        )
+        val result = getPrimitiveForgery(property.returnType, Double::class, forge, forgery)
+        return result ?: uninjectable(property)
     }
 
-    private fun processPropertyWithStringForgery(
+    private fun getStringForgery(
         forge: Forge,
         annotation: StringForgery,
-        property: KMutableProperty<*>,
-        target: Any
-    ) {
+        property: KMutableProperty<*>
+    ): Any {
         val forgery: Forge.() -> String = when (annotation.value) {
             StringForgeryType.ALPHABETICAL -> { -> anAlphabeticalString(annotation.case) }
             StringForgeryType.ALPHA_NUMERICAL -> { -> anAlphaNumericalString(annotation.case) }
@@ -274,50 +268,71 @@ class DefaultForgeryInjector : ForgeryInjector {
             StringForgeryType.ASCII -> { -> anAsciiString() }
             StringForgeryType.ASCII_EXTENDED -> { -> anExtendedAsciiString() }
         }
-        property.setter.invokePrivate(
-            target,
-            forgePrimitiveProperty(property.returnType, String::class, forge, forgery)
-        )
+        val result = getPrimitiveForgery(property.returnType, String::class, forge, forgery)
+        return result ?: uninjectable(property)
     }
 
-    private fun processPropertyWithRegexForgery(
+    private fun getRegexForgery(
         forge: Forge,
         annotation: RegexForgery,
-        property: KMutableProperty<*>,
-        target: Any
-    ) {
+        property: KMutableProperty<*>
+    ): Any {
         val forgery: Forge.() -> String = { aStringMatching(annotation.value) }
-        property.setter.invokePrivate(
-            target,
-            forgePrimitiveProperty(property.returnType, String::class, forge, forgery)
-        )
+        val result = getPrimitiveForgery(property.returnType, String::class, forge, forgery)
+        return result ?: uninjectable(property)
     }
 
-    private fun processPropertyWithAdvancedForgery(
+    private fun getAdvancedForgery(
         forge: Forge,
         annotation: AdvancedForgery,
-        property: KMutableProperty<*>,
-        target: Any
-    ) {
-        if (annotation.string.isNotEmpty()) {
+        property: KMutableProperty<*>
+    ): Any? {
+        return if (annotation.string.isNotEmpty()) {
             val usingAnnotation = forge.anElementFrom(*annotation.string)
-            processPropertyWithStringForgery(forge, usingAnnotation, property, target)
+            getStringForgery(forge, usingAnnotation, property)
         } else if (annotation.int.isNotEmpty()) {
             val usingAnnotation = forge.anElementFrom(*annotation.int)
-            processPropertyWithIntForgery(forge, usingAnnotation, property, target)
+            getIntForgery(forge, usingAnnotation, property)
         } else if (annotation.long.isNotEmpty()) {
             val usingAnnotation = forge.anElementFrom(*annotation.long)
-            processPropertyWithLongForgery(forge, usingAnnotation, property, target)
+            getLongForgery(forge, usingAnnotation, property)
         } else if (annotation.float.isNotEmpty()) {
             val usingAnnotation = forge.anElementFrom(*annotation.float)
-            processPropertyWithFloatForgery(forge, usingAnnotation, property, target)
+            getFloatForgery(forge, usingAnnotation, property)
         } else if (annotation.double.isNotEmpty()) {
             val usingAnnotation = forge.anElementFrom(*annotation.double)
-            processPropertyWithDoubleForgery(forge, usingAnnotation, property, target)
+            getDoubleForgery(forge, usingAnnotation, property)
+        } else if (annotation.map.isNotEmpty()) {
+            val usingAnnotation = forge.anElementFrom(*annotation.map)
+            getMapForgery(forge, usingAnnotation, property)
+        } else {
+            getForgery(forge, property)
         }
     }
 
-    private fun forgeProperty(type: KType, forge: Forge): Any? {
+    private fun getMapForgery(
+        forge: Forge,
+        annotation: MapForgery,
+        property: KMutableProperty<*>
+    ): Map<Any?, Any?> {
+        val mapType = property.returnType
+        check(mapType.classifier in knownMaps)
+
+        val keyAnnotation = annotation.key
+        val keyType = mapType.arguments[0].type ?: uninjectable(property)
+        val keyProperty = KMapProperty<Any>(keyType)
+        val valueAnnotation = annotation.value
+        val valueType = mapType.arguments[1].type ?: uninjectable(property)
+        val valueProperty = KMapProperty<Any>(valueType)
+
+        return forge.aMap {
+            val forgedKey = getAdvancedForgery(forge, keyAnnotation, keyProperty)
+            val forgedValue = getAdvancedForgery(forge, valueAnnotation, valueProperty)
+            forgedKey to forgedValue
+        }
+    }
+
+    private fun getGenericForgery(type: KType, forge: Forge): Any? {
         val arguments = type.arguments
         val classifier = type.classifier
         if (classifier !is KClass<*>) return null
@@ -325,22 +340,22 @@ class DefaultForgeryInjector : ForgeryInjector {
         return if (arguments.isEmpty()) {
             forge.getForgery(classifier.java)
         } else {
-            forgeParameterizedProperty(forge, arguments, classifier)
+            getParameterizedForgery(forge, arguments, classifier)
         }
     }
 
     @Suppress("UnsafeCallOnNullableType")
-    private fun forgeParameterizedProperty(
+    private fun getParameterizedForgery(
         forge: Forge,
         arguments: List<KTypeProjection>,
         classifier: KClass<*>
     ): Any? {
         return when (classifier) {
-            in knownLists -> forge.aList { forgeProperty(arguments[0].type!!, forge) }
-            in knownSets -> forge.aList { forgeProperty(arguments[0].type!!, forge) }.toSet()
+            in knownLists -> forge.aList { getGenericForgery(arguments[0].type!!, forge) }
+            in knownSets -> forge.aList { getGenericForgery(arguments[0].type!!, forge) }.toSet()
             in knownMaps -> forge.aMap {
-                val key = forgeProperty(arguments[0].type!!, forge)
-                val value = forgeProperty(arguments[1].type!!, forge)
+                val key = getGenericForgery(arguments[0].type!!, forge)
+                val value = getGenericForgery(arguments[1].type!!, forge)
                 key to value
             }
             else -> {
@@ -349,7 +364,7 @@ class DefaultForgeryInjector : ForgeryInjector {
         }
     }
 
-    private fun forgePrimitiveProperty(
+    private fun getPrimitiveForgery(
         type: KType,
         klass: KClass<*>,
         forge: Forge,
@@ -366,12 +381,12 @@ class DefaultForgeryInjector : ForgeryInjector {
                 throw IllegalStateException("Unable to forge primitive $klass on proprety with type $type")
             }
         } else {
-            forgeParameterizedPrimitiveProperty(arguments, classifier, klass, forge, forgery)
+            getParameterizedPrimitiveForgery(arguments, classifier, klass, forge, forgery)
         }
     }
 
     @Suppress("UnsafeCallOnNullableType")
-    private fun forgeParameterizedPrimitiveProperty(
+    private fun getParameterizedPrimitiveForgery(
         arguments: List<KTypeProjection>,
         classifier: KClass<*>,
         klass: KClass<*>,
@@ -379,11 +394,52 @@ class DefaultForgeryInjector : ForgeryInjector {
         forgery: Forge.() -> Any?
     ): Any? {
         return when (classifier) {
-            in knownLists -> forge.aList { forgePrimitiveProperty(arguments[0].type!!, klass, forge, forgery) }
-            in knownSets -> forge.aList { forgePrimitiveProperty(arguments[0].type!!, klass, forge, forgery) }.toSet()
-            else -> {
-                null
-            }
+            in knownLists -> forge.aList { getPrimitiveForgery(arguments[0].type!!, klass, forge, forgery) }
+            in knownSets -> forge.aList { getPrimitiveForgery(arguments[0].type!!, klass, forge, forgery) }.toSet()
+            else -> null
+        }
+    }
+
+    private fun uninjectable(property: KMutableProperty<*>): Nothing {
+        throw ForgeryInjectorException("Unable to forge $property.")
+    }
+
+    private class KMapProperty<R>(
+        override val returnType: KType
+    ) : KMutableProperty<R> {
+        override val annotations: List<Annotation>
+            get() = TODO("Not yet implemented")
+        override val getter: KProperty.Getter<R>
+            get() = TODO("Not yet implemented")
+        override val isAbstract: Boolean
+            get() = TODO("Not yet implemented")
+        override val isConst: Boolean
+            get() = TODO("Not yet implemented")
+        override val isFinal: Boolean
+            get() = TODO("Not yet implemented")
+        override val isLateinit: Boolean
+            get() = TODO("Not yet implemented")
+        override val isOpen: Boolean
+            get() = TODO("Not yet implemented")
+        override val isSuspend: Boolean
+            get() = TODO("Not yet implemented")
+        override val name: String
+            get() = TODO("Not yet implemented")
+        override val parameters: List<KParameter>
+            get() = TODO("Not yet implemented")
+        override val setter: KMutableProperty.Setter<R>
+            get() = TODO("Not yet implemented")
+        override val typeParameters: List<KTypeParameter>
+            get() = TODO("Not yet implemented")
+        override val visibility: KVisibility?
+            get() = TODO("Not yet implemented")
+
+        override fun call(vararg args: Any?): R {
+            TODO("Not yet implemented")
+        }
+
+        override fun callBy(args: Map<KParameter, Any?>): R {
+            TODO("Not yet implemented")
         }
     }