+++ inject/src/test/kotlin/fr/xgouchet/elmyr/inject/dummy/KotlinInjectedAdvanced.kt
@@ -81,4 +81,7 @@ class KotlinInjectedAdvanced {
         ]
     )
     var publicMultipleMeansDouble: Double = 0.0
+
+    @AdvancedForgery
+    lateinit var publicGeneric: Foo
 }